// src/app/shared/types/api-types/bidt-df-type-dto.interface.ts
// //////////////////////////////////////////////////////////////////////////
// DO NOT EDIT THIS FILE
//
// It is automatically generated by generate-dtos.js
// //////////////////////////////////////////////////////////////////////////

import { BidtDeliveryFolderDTO } from './bidt-delivery-folder-dto.interface';
import { BidtDfTypeIdDTO } from './bidt-df-type-id-dto.interface';
import { BidtDfTypeOperationDTO } from './bidt-df-type-operation-dto.interface';

export interface BidtDfTypeDTO extends BidtDfTypeIdDTO {
  bidtDeliveryFolders?: BidtDeliveryFolderDTO[];
  bidtDfTypeOperations?: BidtDfTypeOperationDTO[];
  dftCategory?: string; // 50
  dftCode?: string; // 50
  dftDesignation?: string; // 100
  transferType?: string;
}